+++ web_messaging/blueprints/billing/views.py
@@ -7,7 +7,9 @@ from web_messaging.blueprints.billing.storage import retrieve_file_from_bucket
 from web_messaging.extensions import mongo
 from flask_pymongo import pymongo
 
+
 billing = Blueprint('billing', __name__, template_folder='templates')
+
 
 @billing.route("/billing/<filename>", methods=['GET'])
 @login_required
@@ -15,6 +17,7 @@ def get_bill(filename):
     """ Return a bill as an attachment """
     path = retrieve_file_from_bucket(filename)
     return send_file(path, as_attachment=True)
+
 
 @billing.route("/billing", methods=['GET'])
 @login_required
@@ -25,11 +28,13 @@ def bills():
     bills = cursor.sort("date", pymongo.ASCENDING) 
     return render_template("billing.html", bills=bills)
 
+
 def create_new_bill(file, billing_date, total_cost_usd):
     """ Create a new Bill object and upload that object on MongoDB"""
     new_bill = Bill(billing_date, int(total_cost_usd), file)
     new_bill.upload_to_gcs()
     mongo.db['billing'].insert_one(new_bill.dict())
+    
     
 @billing.route("/upload-bill", methods = ['POST', 'GET'])
 @login_required